Update readme for Vue CLI 3
With the change to Vue CLI 3 the vue documentation no longer works.
To update users need to import path in addition to
purgecss-webpack-plugin and then set vue.config.js as shown in the
commit.

After testing a repo with Bootstrap added and no classes called, the
file size reduction is nearly the same as what is referenced in the
original ~3.4kb